chore: extract trading engine and add DDD scaffolding
