+++ docs/typing-migration/checklists/phase-08.md
@@ -3,7 +3,7 @@
 - [ ] Refactor `TODO_COMMENTS.md`
 - [ ] Refactor `the_alchemiser/application/tracking/integration.py`
 - [ ] Refactor `the_alchemiser/application/tracking/strategy_order_tracker.py`
-- [ ] Refactor `the_alchemiser/application/trading/trading_engine.py`
+- [ ] Refactor `the_alchemiser/application/engine_service.py`
 - [ ] Refactor `the_alchemiser/application/types.py`
 - [ ] Refactor `the_alchemiser/execution/account_service.py`
 - [ ] Refactor `the_alchemiser/interface/email/templates/portfolio.py`
