+++ reports/mypy_after.txt
@@ -1,0 +1,1 @@
+Success: no issues found in 188 source files
