+++ the_alchemiser/domain/services/__init__.py
@@ -1,0 +1,1 @@
+"""Domain service layer containing pure business policies."""
